refactor: rename package view to io for clearer I/O responsibility
