+++ src/main/java/lotto/io/OutputView.java
@@ -1,4 +1,4 @@
-package lotto.view;
+package lotto.io;
 
 import java.math.BigDecimal;
 import java.text.DecimalFormat;
