+++ src/test/java/lotto/io/InputValidatorTest.java
@@ -1,4 +1,4 @@
-package lotto.view;
+package lotto.io;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
